Accept key sizes of 128, 192, or 256 as input to aes_cfb_128_crypt
Despite the confusion caused by the name, aes_cfb_128_crypt can use
key lengths of 128, 192, or 256. The 128 in the function name refers
to the fixed block size of the cipher.  Change the function to accept
and use a key of length 128, 192, or 256. Also add testing using the
NIST test vectors for the additional key lengths to ensure the changes
function properly.